use structs and interfaces to make the result closer to the original
Although switching to a map allowed for a non-integer driven strategy,
we lost something in the process - our output is now in an undefined
order, whereas before it was sorted by ascending k. Though we now have a
strategy that is not dependent on k, it would still be nice to have the
stayAtK strategy results grouped together and meaningfully ordered.
Turning strategy into an interface which implements fmt.Stringer, we can
get rid of the need for map keys (the human-readable label is now
provided by the Strategy itself) and go back to using an array, simply
appending an instance of the Random implementation at the end.